Rename cookiecutter.logging module to log
